Inclusion of analysis data
The Airbnb_seattle_2017_analysis.py file contains the analysis for all three questions, comments on those and a detailed and fully commented guidiance through the data preperation and visualization of all three questions.

The Readme is updated and contains the relevant information that helps to work with the code

The necessary data is stored in the rep.